fix: https multiaddr support in constructor
Passing HTTPS multiaddr as a string to constructor did not work,
protocol was ignored and unencrypted HTTP was used.

This commit improves parsing of multiaddr constructor parameter
and adds tests to ensure HTTPS is respected in all scenarios.

It also updates iso-stream-http to v0.1.2
which includes fix for "https does not exist"

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>