build: change min node version to 8.3.0
Also adjust Travis CI config to perform builds and tests with node 8.

`npm install`, `npm run build`, and `npm run test` all succeeded locally with
node version `8.3.0`.

The reason for version `8.3.0` instead of `8.0.0` is that `8.3.0` introduced
support for object rest/spread, which is used by at least one of this package's
devDependencies (`go-ipfs-dep`). It may be the case that at runtime older
versions of node would work correctly with the built package, but it seems
simpler to settle on the minimum supported version being a version that can
build, test, and use the package.

Closes #983.