fix: increase default timeout and respect value passed to `ky.extend` (#1130)
* fix: disable timeout if not set

Some of our operations take a really long time, if we don't set a
timeout for `ky` we get the default of 10 seconds. This PR sets
the timeout to `false` if one is not explicitly passed which
disables it.

Nb. I had to add the default to `false` to every invocation. Looking
at the code it should be enough to do it in `src/lib/configure.js`
but it doesn't seem to be.

* fix: use `ignoreUndefined` merge-options option

* chore: update bundle size

* chore: remove git url